Try of new if check

+++ sudo/module.py
@@ -127,6 +127,12 @@ class Sudo(commands.Cog):
             await ctx.reply(_(ctx, "Message must be shorter than 2000 characters."))
             return
 
+        if message is not discord.message:
+            await ctx.reply(
+                _(ctx, "Message does not exist!")
+            )
+            return
+
         await dc_message.edit(content=message)
         await utils.discord.delete_message(ctx.message)
         await ctx.send(
